Merge `CODE_SIGN_STYLE` with `ProvisioningStyle` (#11)
Apparently, Xcode 9 introduced a new build setting `CODE_SIGN_STYLE`,
which does the same as `ProvisioningStyle` target attribute setting
introduced on Xcode 8. Additionally, on Xcode 9.2 it appears to be the
only value Xcode considers when building, so we can assume it will be
the default going forward.

Given that both options might be required (we don't know which Xcode is
being used to build), for now we simply apply the setting on both
places.

Also, since the way forward seems to be using `CODE_SIGN_STYLE`, the
parameter `:provisioning_style` was renamed to `:code_sign_style`
(which IMO now is more clear about what it does).